big fat refactoring begins
