+++ src/FluffySpoon.AspNet.LetsEncrypt/LetsEncryptRenewalService.cs
@@ -1,70 +1,58 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Threading.Tasks;
-using Certes;
-using Certes.Acme;
-using Certes.Acme.Resource;
-using FluffySpoon.AspNet.LetsEncrypt.Exceptions;
-using FluffySpoon.AspNet.LetsEncrypt.Persistence;
-using FluffySpoon.AspNet.LetsEncrypt.Persistence.Models;
 using Microsoft.Extensions.Logging;
 
 namespace FluffySpoon.AspNet.LetsEncrypt
 {
 	public class LetsEncryptRenewalService : ILetsEncryptRenewalService
 	{
-		public const string CertificateFriendlyName = "FluffySpoonAspNetLetsEncryptCertificate";
-
+		private readonly ICertificateRenewal _logic;
 		private readonly IEnumerable<ICertificateRenewalLifecycleHook> _lifecycleHooks;
-		private readonly IPersistenceService _persistenceService;
 		private readonly ILogger<ILetsEncryptRenewalService> _logger;
-
-		private readonly LetsEncryptOptions _options;
 		private readonly SemaphoreSlim _semaphoreSlim;
-
-		private IAcmeContext acme;
+		private readonly LetsEncryptOptions _options;
 
 		private Timer _timer;
 
-		public static X509Certificate2 Certificate { get; private set; }
-
-		public Uri LetsEncryptUri => _options.UseStaging
-			? WellKnownServers.LetsEncryptStagingV2
-			: WellKnownServers.LetsEncryptV2;
-
 		public LetsEncryptRenewalService(
+			ICertificateRenewal logic,
 			IEnumerable<ICertificateRenewalLifecycleHook> lifecycleHooks,
-			IPersistenceService persistenceService,
 			ILogger<ILetsEncryptRenewalService> logger,
 			LetsEncryptOptions options)
 		{
+			_logic = logic;
 			_lifecycleHooks = lifecycleHooks;
-			_persistenceService = persistenceService;
 			_logger = logger;
 			_options = options;
-
 			_semaphoreSlim = new SemaphoreSlim(1);
 		}
 
-		public void Dispose()
-		{
-			_timer?.Dispose();
-		}
-
+		public static X509Certificate2 Certificate { get; private set; }
+		
+		public Uri LetsEncryptUri => _logic.LetsEncryptUri;
+		
 		public async Task StartAsync(CancellationToken cancellationToken)
 		{
 			if (_options.TimeAfterIssueDateBeforeRenewal == null && _options.TimeUntilExpiryBeforeRenewal == null)
+			{
 				throw new InvalidOperationException(
-					"Neither TimeAfterIssueDateBeforeRenewal nor TimeUntilExpiryBeforeRenewal have been set, which means that the LetsEncrypt certificate will never renew.");
+					"Neither TimeAfterIssueDateBeforeRenewal nor TimeUntilExpiryBeforeRenewal have been set," +
+					" which means that the LetsEncrypt certificate will never renew.");
+			}
 
 			foreach (var lifecycleHook in _lifecycleHooks)
+			{
 				await lifecycleHook.OnStartAsync();
+			}
 
-			_timer = new Timer(async (state) => await RunOnceWithErrorHandlingAsync(), null, TimeSpan.Zero, TimeSpan.FromHours(1));
+			_timer = new Timer(
+				callback: async state => await RunOnceWithErrorHandlingAsync(), 
+				state: null,
+				dueTime: TimeSpan.Zero, 
+				period: TimeSpan.FromHours(1));
 		}
 
 		public async Task StopAsync(CancellationToken cancellationToken)
@@ -76,43 +64,6 @@ namespace FluffySpoon.AspNet.LetsEncrypt
 				await lifecycleHook.OnStopAsync();
 		}
 
-		private async Task<bool> TryRetrievingValidPersistedSiteCertificateAsync()
-		{
-			if (Certificate != null)
-				return false;
-
-			var certificate = await _persistenceService.GetPersistedSiteCertificateAsync();
-			if (certificate == null)
-				return false;
-
-			Certificate = certificate;
-
-			if (IsCertificateValid)
-			{
-				_logger.LogInformation("A persisted non-expired LetsEncrypt certificate was found and will be used.");
-				return true;
-			}
-
-			_logger.LogInformation("A persisted but expired LetsEncrypt certificate was found and will be renewed.");
-
-			return false;
-		}
-
-		private async Task RunOnceWithErrorHandlingAsync()
-		{
-			try
-			{
-				await RunOnceAsync();
-				_timer.Change(Timeout.InfiniteTimeSpan, TimeSpan.FromHours(1));
-			}
-			catch (Exception e) when (_options.RenewalFailMode != RenewalFailMode.Unhandled)
-			{
-				_logger.LogWarning(e, $"Exception occured renewing certificates: '{e.Message}.'");
-				if (_options.RenewalFailMode == RenewalFailMode.LogAndRetry)
-					_timer.Change(Timeout.InfiniteTimeSpan, TimeSpan.FromMinutes(1));
-			}
-		}
-
 		public async Task RunOnceAsync()
 		{
 			if (_semaphoreSlim.CurrentCount == 0)
@@ -120,23 +71,9 @@ namespace FluffySpoon.AspNet.LetsEncrypt
 
 			await _semaphoreSlim.WaitAsync();
 
-			_logger.LogInformation("Checking to see if in-memory LetsEncrypt certificate needs renewal.");
-
 			try
 			{
-				if (IsCertificateValid)
-					return;
-
-				_logger.LogInformation("Checking to see if existing LetsEncrypt certificate has been persisted and is valid.");
-
-				var alreadyHasValidCertificate = await TryRetrievingValidPersistedSiteCertificateAsync();
-				if (alreadyHasValidCertificate)
-					return;
-
-				await AuthenticateAsync();
-
-				var domains = _options.Domains?.ToArray() ?? Array.Empty<string>();
-				await AcquireNewCertificateForDomains(domains);
+				Certificate = await _logic.RenewCertificateIfNeeded(Certificate);
 			}
 			catch (Exception ex)
 			{
@@ -151,168 +88,24 @@ namespace FluffySpoon.AspNet.LetsEncrypt
 			}
 		}
 
-		private async Task AcquireNewCertificateForDomains(string[] domains)
+		private async Task RunOnceWithErrorHandlingAsync()
 		{
-			_logger.LogInformation("Ordering LetsEncrypt certificate for domains {0}.", new object[] { domains });
-
-			var order = await acme.NewOrder(domains);
-			await ValidateOrderAsync(domains, order);
-
-			var certificateBytes = await AcquireCertificateBytesFromOrderAsync(order);
-			if (certificateBytes == null)
-			{
-				throw new InvalidOperationException("The certificate from the order was null.");
-			}
-
-			await _persistenceService.PersistSiteCertificateAsync(certificateBytes);
-
-			Certificate = new X509Certificate2(certificateBytes, nameof(FluffySpoon));
-		}
-
-		private async Task AuthenticateAsync()
-		{
-			if (acme != null)
-				return;
-
-			var existingAccountKey = await _persistenceService.GetPersistedAccountCertificateAsync();
-			if (existingAccountKey != null)
-			{
-				await UseExistingLetsEncryptAccount(existingAccountKey);
-			}
-			else
-			{
-				await CreateNewLetsEncryptAccount();
-			}
-		}
-
-		private async Task UseExistingLetsEncryptAccount(IKey existingAccountKey)
-		{
-			_logger.LogDebug("Using existing LetsEncrypt account.");
-
-			acme = new AcmeContext(LetsEncryptUri, existingAccountKey);
-			await acme.Account();
-		}
-
-		private async Task CreateNewLetsEncryptAccount()
-		{
-			_logger.LogDebug("Creating LetsEncrypt account with email {0}.", _options.Email);
-
-			acme = new AcmeContext(LetsEncryptUri);
-			await acme.NewAccount(_options.Email, true);
-
-			await _persistenceService.PersistAccountCertificateAsync(acme.AccountKey);
-		}
-
-		private async Task<byte[]> AcquireCertificateBytesFromOrderAsync(IOrderContext order)
-		{
-			_logger.LogInformation("Acquiring certificate through signing request.");
-
-			var privateKey = KeyFactory.NewKey(KeyAlgorithm.ES256);
-			var certificateChain = await order.Generate(
-				_options.CertificateSigningRequest,
-				privateKey);
-
-			var pfxBuilder = certificateChain.ToPfx(privateKey);
-			pfxBuilder.FullChain = true;
-
-			var pfxBytes = pfxBuilder.Build(CertificateFriendlyName, nameof(FluffySpoon));
-
-			_logger.LogInformation("Certificate acquired.");
-
-			return pfxBytes;
-		}
-
-		private async Task ValidateOrderAsync(string[] domains, IOrderContext order)
-		{
-			var allAuthorizations = await order.Authorizations();
-			var challengeContextTasks = new List<Task<IChallengeContext>>();
-
-			challengeContextTasks.AddRange(allAuthorizations.Select(x => x.Http()));
-
-			var challengeContexts = await Task.WhenAll(challengeContextTasks);
-			var nonNullChallengeContexts = challengeContexts.Where(x => x != null);
-
-			_logger.LogInformation("Validating all pending order authorizations.");
-
-			var challengeDtos = nonNullChallengeContexts.Select(x => new ChallengeDto()
-			{
-				Token = x.Type == ChallengeTypes.Dns01 ? 
-                    acme.AccountKey.DnsTxt(x.Token) : 
-                    x.Token,
-				Response = x.KeyAuthz,
-				Domains = domains
-			}).ToArray();
-
-			await _persistenceService.PersistChallengesAsync(challengeDtos);
-
 			try
 			{
-				var challengeValidationResponses = await ValidateChallengesAsync(nonNullChallengeContexts);
-				var nonNullChallengeValidationResponses = challengeValidationResponses.Where(x => x != null).ToArray();
-
-				if (challengeValidationResponses.Length > nonNullChallengeValidationResponses.Length)
-					_logger.LogWarning("Some challenge responses were null.");
-							   
-				await _persistenceService.DeleteChallengesAsync(challengeDtos);
-
-				var challengeExceptions = nonNullChallengeValidationResponses
-					.Where(x => x.Status == ChallengeStatus.Invalid)
-					.Select(x => new Exception($"{x.Error?.Type ?? "errortype null"}: {x.Error?.Detail ?? "null errordetails"} (challenge type {x.Type ?? "null"})"))
-					.ToArray();
-
-				if (challengeExceptions.Length > 0) 
-					throw new OrderInvalidException(
-						"One or more LetsEncrypt orders were invalid. Make sure that LetsEncrypt can contact the domain you are trying to request an SSL certificate for, in order to verify it.",
-						new AggregateException(challengeExceptions));
-                
+				await RunOnceAsync();
+				_timer?.Change(Timeout.InfiniteTimeSpan, TimeSpan.FromHours(1));
 			}
-			finally
+			catch (Exception e) when (_options.RenewalFailMode != RenewalFailMode.Unhandled)
 			{
-				await _persistenceService.DeleteChallengesAsync(challengeDtos);
+				_logger.LogWarning(e, $"Exception occured renewing certificates: '{e.Message}.'");
+				if (_options.RenewalFailMode == RenewalFailMode.LogAndRetry)
+					_timer?.Change(Timeout.InfiniteTimeSpan, TimeSpan.FromMinutes(1));
 			}
 		}
 
-		private async Task<Challenge[]> ValidateChallengesAsync(IEnumerable<IChallengeContext> challengeContexts)
+		public void Dispose()
 		{
-			var challenges = await Task.WhenAll(
-								challengeContexts.Select(x => x.Validate()));
-
-			while (true)
-			{
-				if (challenges.Any(x => x.Status == ChallengeStatus.Valid) || challenges.All(x => x.Status == ChallengeStatus.Invalid))
-					break;
-                
-                await Task.Delay(1000);
-				challenges = await Task.WhenAll(challengeContexts.Select(x => x.Resource()));
-			}
-
-			return challenges;
+			_timer?.Dispose();
 		}
-
-		private bool IsCertificateValid
-		{
-			get
-			{
-				try
-				{
-					if (Certificate == null)
-						return false;
-					else if (_options.TimeUntilExpiryBeforeRenewal != null && Certificate.NotAfter - DateTime.Now < _options.TimeUntilExpiryBeforeRenewal)
-						return false;
-					else if (_options.TimeAfterIssueDateBeforeRenewal != null && DateTime.Now - Certificate.NotBefore > _options.TimeAfterIssueDateBeforeRenewal)
-						return false;
-					else if (Certificate.NotBefore > DateTime.Now || Certificate.NotAfter < DateTime.Now)
-						return false;
-					else
-						return true;
-				}
-				catch (CryptographicException exc)
-				{
-					_logger.LogError(exc, "Exception occured during certificate validation");
-					return false;
-				}
-			}
-		}
-
 	}
 }
